+++ src/commands.ts
@@ -68,6 +68,7 @@ import { ClearBlame } from "./commands/blame/clearBlame";
 import { EnableBlame } from "./commands/blame/enableBlame";
 import { DisableBlame } from "./commands/blame/disableBlame";
 import { UntrackedInfo } from "./commands/blame/untrackedInfo";
+import { ApplyRecommendedSettings } from "./commands/applyRecommendedSettings";
 
 export function registerCommands(
   sourceControlManager: SourceControlManager,
@@ -139,4 +140,5 @@ export function registerCommands(
   disposables.push(new EnableBlame());
   disposables.push(new DisableBlame());
   disposables.push(new UntrackedInfo());
+  disposables.push(new ApplyRecommendedSettings());
 }